backend/services/llm_orchestrator/requirements.txt
# Core Agentic Libraries
langchain>=0.2.0
langgraph>=0.1.0
langchain-community
langchain-core

# Azure LLM & SDK
azure-ai-inference>=1.0.0b2
azure-identity

# Web Framework & Server
fastapi
uvicorn

# Memory
redis

# Utilities
python-dotenv
pydantic
tenacity
aiohttp
chromadb
httpx

# Async HTTP
httpx
aiohttp

# Memory & caching
chromadb

# Utilities
tenacity  # For retries 